+++ .circleci/config.yml
@@ -43,7 +43,7 @@ jobs:
                 command: cargo build --workspace --jobs 2
             - run:
                 name: Test Trin workspace
-                command: cargo test --workspace --jobs 2 -- --nocapture
+                command: cargo test --workspace --jobs 2
   win-build:
     description: |
       Build the crate for windows.
