+++ newsfragments/349.fixed.md
@@ -1,1 +1,1 @@
-Encode content items send over uTP with variable length unsigned integer (varint) prefixing each content item.
+Encode content items sent over uTP with variable length unsigned integer (varint) prefixing each content item.
