+++ trin-core/src/portalnet/overlay.rs
@@ -1,6 +1,4 @@
 use anyhow::anyhow;
-use bytes::{Buf, BytesMut};
-use std::io::Read;
 use std::{
     collections::{BTreeMap, HashMap, HashSet},
     fmt::{Debug, Display},
@@ -25,6 +23,7 @@ use crate::portalnet::{
     },
 };
 
+use crate::utils::portal_wire;
 use crate::{
     portalnet::types::{content_key::RawContentKey, messages::ByteList, metric::XorMetric},
     types::validation::Validator,
@@ -241,7 +240,7 @@ where
         content_keys: Vec<RawContentKey>,
         payload: UtpPayload,
     ) -> anyhow::Result<()> {
-        let content_values = read_content_values(payload)?;
+        let content_values = portal_wire::decode_content_payload(payload)?;
 
         // Accepted content keys len should match content value len
         if content_keys.len() != content_values.len() {
@@ -657,48 +656,6 @@ fn log2_random_enrs(enrs: Vec<Enr>) -> anyhow::Result<Vec<Enr>> {
     Ok(random_enrs)
 }
 
-/// Decode content values from uTP payload. All content values are encoded with a LEB128 varint prefix
-/// which indicates the length in bytes of the consecutive content item.
-fn read_content_values(payload: Vec<u8>) -> anyhow::Result<Vec<ByteList>> {
-    let mut payload = BytesMut::from(&payload[..]).reader();
-
-    let mut content_values: Vec<ByteList> = Vec::new();
-
-    // Read LEB128 encoded index and content items until all payload bytes are consumed
-    while !payload.get_ref().is_empty() {
-        // Read LEB128 index
-        let (bytes_to_read, varint) = read_varint(&mut payload.get_ref())?;
-        let mut buf = vec![0u8; bytes_to_read];
-        payload
-            .read_exact(&mut buf)
-            .map_err(|err| anyhow!("Error reading varint index: {err}"))?;
-
-        // Read the content item
-        let mut buf = vec![0u8; varint as usize];
-        payload
-            .read_exact(&mut buf)
-            .map_err(|err| anyhow!("Error reading content item: {err}"))?;
-        content_values.push(buf.into());
-    }
-    Ok(content_values)
-}
-
-/// Try to read up to five LEB128 bytes (The maximum content size allowed for this application is limited to `uint32`).
-fn read_varint(buf: &[u8]) -> anyhow::Result<(usize, u32)> {
-    for i in 1..6 {
-        match leb128::read::unsigned(&mut &buf[0..i]) {
-            Ok(varint) => {
-                let varint = u32::try_from(varint).map_err(|_| {
-                    anyhow!("Exceed maximum allowed varint value of u32 bytes size")
-                })?;
-                return Ok((i, varint));
-            }
-            Err(_) => continue,
-        }
-    }
-    Err(anyhow!("Unable to read varint index"))
-}
-
 fn validate_find_nodes_distances(distances: &Vec<u16>) -> Result<(), OverlayRequestError> {
     if distances.len() == 0 {
         return Err(OverlayRequestError::InvalidRequest(
@@ -771,68 +728,5 @@ mod test {
     fn test_log2_random_enrs_empty_input() {
         let all_nodes = vec![];
         log2_random_enrs(all_nodes).unwrap();
-    }
-
-    #[rstest]
-    #[case(u8::MIN as u32)]
-    #[case(u8::MAX as u32)]
-    #[case(u16::MAX as u32)]
-    #[case(u32::MAX)]
-    fn test_read_varint(#[case] varint: u32) {
-        let mut buf = [0; 1024];
-        let mut writable = &mut buf[..];
-
-        let bytes_written = leb128::write::unsigned(&mut writable, varint.into()).unwrap();
-
-        let (bytes_read, varint_result) = read_varint(&mut buf[..]).unwrap();
-
-        assert_eq!(bytes_read, bytes_written);
-        assert_eq!(varint_result, varint);
-    }
-
-    #[test]
-    #[should_panic(expected = "Unable to read varint index")]
-    fn test_read_varint_max_read_bytes() {
-        let mut buf = [0; 1024];
-        let mut writable = &mut buf[..];
-        let varint = u64::MAX;
-
-        leb128::write::unsigned(&mut writable, varint).unwrap();
-
-        read_varint(&mut buf[..]).unwrap();
-    }
-
-    #[test]
-    #[should_panic(expected = "Exceed maximum allowed varint value of u32 bytes size")]
-    fn test_read_varint_max_varint_size() {
-        let mut buf = [0; 1024];
-        let mut writable = &mut buf[..];
-        let varint = u32::MAX as usize + 1;
-
-        leb128::write::unsigned(&mut writable, varint as u64).unwrap();
-
-        read_varint(&mut buf[..]).unwrap();
-    }
-
-    #[test]
-    fn test_read_content_values() {
-        let hex_payload = "03010101020101";
-        let payload = hex::decode(hex_payload).unwrap();
-        let content_values = read_content_values(payload).unwrap();
-
-        let expected_content_values = vec![
-            ByteList::new(vec![1, 1, 1]).unwrap(),
-            ByteList::new(vec![1, 1]).unwrap(),
-        ];
-
-        assert_eq!(content_values, expected_content_values);
-    }
-
-    #[test]
-    #[should_panic(expected = "Error reading content item: failed to fill whole buffer")]
-    fn test_read_content_values_corrupted_payload() {
-        let hex_payload = "030101010201";
-        let payload = hex::decode(hex_payload).unwrap();
-        read_content_values(payload).unwrap();
     }
 }
