+++ trin-core/src/utils/mod.rs
@@ -3,3 +3,4 @@ pub mod bytes;
 pub mod db;
 pub mod infura;
 pub mod node_id;
+pub mod portal_wire;
